100s of hacks , bidi on CH seems to work, but  weight function not correct fue to junction

+++ core/src/main/java/com/graphhopper/routing/AbstractRoutingAlgorithm.java
@@ -103,7 +103,7 @@ public abstract class AbstractRoutingAlgorithm implements RoutingAlgorithm
         return additionalEdgeFilter == null || additionalEdgeFilter.accept(iter);
     }
 
-    protected void updateBestPath( EdgeIteratorState edgeState, EdgeEntry bestEdgeEntry, int key )
+    protected void updateBestPath( EdgeIteratorState edgeState, EdgeEntry bestEdgeEntry, long key )
     {
     }
